pyproject.toml
[project]
name = "pj-ogun"
version = "0.1.0"
description = "Logistics & Field Operations Simulation Platform"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",
    "simpy>=4.1",
    "networkx>=3.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "streamlit>=1.30",
    "plotly>=5.18",
    "streamlit-flow-component>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
ogun = "pj_ogun.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=pj_ogun"
